bug in consumer and cat_service fixed
