+++ src/configs/service_config.py
@@ -3,8 +3,8 @@ import os
 from dotenv import load_dotenv
 
 from dataclasses import dataclass
-from src.telegram_bot.configs.bot_base_configs import S3ClientConfig
 
+from src.services.configs.services_base_configs import S3ClientConfig, RedisClientConfig
 
 load_dotenv()
 
@@ -13,9 +13,13 @@ class ServiceConfig:
     mongoDB_url: str
     bot_token: str
     s3_client_config: S3ClientConfig
+    redis_client_config: RedisClientConfig
     models_path: str
     local_models_path: str
     embedding_size: int
+    answer_time_delay: int  ## dely time for bot loader (one cat  in 5 sec)
+    ans_check_frequency: int  ## how often do we check for answers in main _nference_loop
+    cats_in_answer: int
 
 
 default_service_config = ServiceConfig(
@@ -26,6 +30,10 @@ default_service_config = ServiceConfig(
                     aws_secret_access_key=os.environ.get('AWS_SECRET_ACCESS_KEY'),
                     local_path=os.environ.get('LOCAL_PATH'),
                 ),
+    redis_client_config=RedisClientConfig(host=os.environ.get("REDIS_URL"),
+                                          port=6379,
+                                          db=0
+    ),
     models_path=os.environ.get('MODEL_PATH'),
     local_models_path=os.environ.get('LOCAL_MODEL_PATH'),
     embedding_size=512,
